advanced

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,5 +1,4 @@
 #include <stdio.h>
-
 /**
  * main - Prints first 50 Fibonacci numbers, starting with 1 and 2,
  *        separated by a comma followed by a space.
@@ -8,7 +7,7 @@
  */
 int main(void)
 {
-	int count;
+	short count;
 	unsigned long fib1 = 0, fib2 = 1, sum;
 
 	for (count = 0; count < 50; count++)
